requirements.txt
# === Base ML/CUDA stack ===
torch==2.3.0+cu118
torchvision==0.18.0+cu118
torchaudio==2.3.0+cu118

# === Core RAG / NLP ===
sentence-transformers>=3.2.1
transformers==4.57.1
huggingface-hub==0.36.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy>=1.14.1
numpy==2.3.3
pandas==2.3.3

# === Vector DB / Embeddings ===
chromadb>=1.2.2
qdrant-client>=1.12.1
onnxruntime==1.23.2
faiss-cpu>=1.8.0

# === API / Web / UI ===
fastapi==0.120.1
uvicorn==0.38.0
streamlit==1.50.0
jinja2==3.1.6
python-dotenv==1.2.1
requests==2.32.5
tqdm==4.67.1

# === PDF / DOC parsers ===
PyPDF2==3.0.1
pymupdf==1.24.9
pdfplumber==0.11.4
python-docx==1.1.2
docx2txt==0.8
unstructured==0.15.13
camelot-py[cv]==0.11.0
tabula-py==2.9.3

# === Utility / Logs / Build ===
coloredlogs==15.0.1
humanfriendly==10.0
typer==0.20.0
filelock==3.19.1
tenacity==9.1.2
rich==14.2.0
watchdog==6.0.0
requests-oauthlib==2.0.0
orjson==3.11.4

# === Optional Dev Tools ===
pytest==8.3.3
build==1.3.0
wheel==0.44.0
